requirements.txt
absl-py==1.4.0
altgraph==0.17.2
anyio==4.2.0
APScheduler==3.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.0
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
cattrs==23.2.3
cchardet==2.1.7
certifi==2022.9.14
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.32
ddddocr==1.4.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.0
dnspython==2.2.1
docstring-to-markdown==0.13
emoji==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.3
fake-useragent==1.2.1
fastjsonschema==2.19.0
filelock==3.8.2
findspark==2.0.1
fiona==1.9.5
Flask==1.0
flatbuffers==22.11.23
fonttools==4.28.4
fst-pso==1.8.1
funcy==1.18
future==0.18.2
FuzzyTM==2.0.5
fuzzywuzzy==0.18.0
gast==0.4.0
gensim==4.3.0
geopandas==0.13.2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.54.0
gunicorn==19.9.0
h11==0.14.0
h5py==3.8.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
image==1.5.33
importlib-metadata==6.6.0
importlib-resources==6.0.1
incremental==22.10.0
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jax==0.4.8
jedi==0.19.1
jedi-language-server==0.41.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-language-pack-zh-CN==4.0.post6
jupyterlab-lsp==5.0.1
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kafka==1.3.5
keras==2.12.0
kiwisolver==1.3.2
libclang==16.0.0
lsprotocol==2023.0.0
lxml==4.9.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==3.0.2
ml-dtypes==0.1.0
mpld3==0.5.10
mpmath==1.2.1
Naked==0.1.32
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
patsy==0.5.5
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
pip-review==1.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.18.0
prettytable==3.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
Protego==0.2.1
protobuf==4.21.9
psutil==5.9.7
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclust==0.2.0
pycparser==2.21
pycryptodome==3.20.0
pydash==7.0.6
PyDispatcher==2.0.6
pydotplus==2.0.2
pyecharts==2.0.3
PyExecJS==1.5.1
pyFUME==0.2.25
pygal==3.0.4
pygls==1.2.1
Pygments==2.17.2
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyLDAvis==3.4.0
pymongo==4.3.2
PyMySQL==1.0.3
pynput==1.7.6
pyOpenSSL==22.1.0
pyparsing==3.0.6
pyproj==3.5.0
pyreadline3==3.4.1
PySocks==1.7.1
pyspark==3.5.0
pyspark-stubs==3.0.0.dev0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.2.1
pywencai==0.12.0
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
redis==4.3.4
referencing==0.32.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
Scrapy==2.7.1
seaborn==0.12.2
selenium==4.6.0
selenium-requests==2.0.3
Send2Trash==1.8.2
service-identity==21.1.0
shapely==2.0.3
shellescape==3.8.1
simpful==2.10.0
simplejson==3.19.1
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.3.0
sniffio==1.3.0
snownlp==0.12.3
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.27
sqlparse==0.4.2
squarify==0.4.3
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
torch==2.0.1
torch-cluster @ file:///D:/%E4%B8%B4%E6%97%B6%E6%96%87%E4%BB%B6/%E6%B5%8F%E8%A7%88%E5%99%A8%E4%B8%8B%E8%BD%BD%E6%96%87%E4%BB%B6/torch_cluster-1.6.3%2Bpt20cpu-cp38-cp38-win_amd64.whl
torch-scatter @ file:///D:/%E4%B8%B4%E6%97%B6%E6%96%87%E4%BB%B6/%E6%B5%8F%E8%A7%88%E5%99%A8%E4%B8%8B%E8%BD%BD%E6%96%87%E4%BB%B6/torch_scatter-2.1.2%2Bpt20cpu-cp38-cp38-win_amd64.whl
torch-sparse @ file:///D:/%E4%B8%B4%E6%97%B6%E6%96%87%E4%BB%B6/%E6%B5%8F%E8%A7%88%E5%99%A8%E4%B8%8B%E8%BD%BD%E6%96%87%E4%BB%B6/torch_sparse-0.6.18%2Bpt20cpu-cp38-cp38-win_amd64.whl
torch-spline-conv @ file:///D:/%E4%B8%B4%E6%97%B6%E6%96%87%E4%BB%B6/%E6%B5%8F%E8%A7%88%E5%99%A8%E4%B8%8B%E8%BD%BD%E6%96%87%E4%BB%B6/torch_spline_conv-1.2.2%2Bpt20cpu-cp38-cp38-win_amd64.whl
torch_geometric==2.4.0
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
traitlets==5.14.0
treelib==1.6.1
trio==0.22.0
trio-websocket==0.9.2
tushare==1.2.89
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.9.0
tzdata==2021.5
tzlocal==5.0.1
ufile==3.2.6
urllib3==1.24.3
w3lib==2.1.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==0.57.0
wencai==0.2.6
Werkzeug==0.15.5
widgetsnbextension==4.0.9
wordcloud==1.8.1
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
zipp==3.15.0
zope.interface==5.5.2
